Add minimalist shim for XH\Mail
This may be useful for plugins which already use XH\Mail for sending
emails, but for whatever reason want to switch using PHPMailer under
the hood, by just switching the namespace.  This should work for some
typical cases, but is not supposed to ever provide full compatibility
with XH\Mail.