Configurable sender address and name
This is likely required for sensing via SMTP, but generally best
practice when sending mail from a Website: always use your own email
address as `From` address.  For a contact form or such, set the email
address of the user as `Reply-To`.

Thus, we always apply the configured address (the name is optional),
and add a system check.